- Added support for SNMP trap logging (server side) - More changes related to 64bit platforms - Fixed bug in event template message expanding

+++ src/server/tools/nxconfig/SummaryPage.cpp
@@ -56,7 +56,7 @@ static void AddParam(TCHAR *pszBuffer, TCHAR *pszText, void *pValue)
 {
    int iLen;
 
-   iLen = _tcslen(pszBuffer);
+   iLen = (int)_tcslen(pszBuffer);
    _sntprintf(&pszBuffer[iLen], BUFFER_SIZE - iLen, pszText, pValue);
 }
